+++ src/utils/displayFilename.ts
@@ -1,5 +1,5 @@
 /**
- * useDisplayFilename Hook
+ * getDisplayFilename Utility
  *
  * Determines which filename to display based on priority:
  * 1. currentFileName (from pattern store)
@@ -8,7 +8,7 @@
  * 4. Empty string
  */
 
-export function useDisplayFilename(options: {
+export function getDisplayFilename(options: {
   currentFileName: string | null;
   localFileName: string;
   resumeFileName: string | null;
